+++ examples/official-storybook/stories/addon-controls.stories.tsx
@@ -5,10 +5,15 @@ export default {
   title: 'Addons/Controls',
   component: Button,
   argTypes: {
-    children: { control: 'text', name: 'Children' },
+    children: { control: 'text', name: 'Children', mapping: { basic: 'BASIC' } },
     type: { control: 'text', name: 'Type' },
     json: { control: 'object', name: 'JSON' },
     imageUrls: { control: { type: 'file', accept: '.png' }, name: 'Image Urls' },
+    label: {
+      name: 'Label',
+      control: { type: 'select', options: ['Plain', 'Bold'] },
+      mapping: { Bold: <b>Bold</b> },
+    },
   },
   parameters: { chromatic: { disable: true } },
 };
@@ -17,7 +22,7 @@ const DEFAULT_NESTED_OBJECT = { a: 4, b: { c: 'hello', d: [1, 2, 3] } };
 
 const Template = (args) => (
   <div>
-    <Button type={args.type}>{args.children}</Button>
+    <Button type={args.type}>{args.label || args.children}</Button>
     {args.json && <pre>{JSON.stringify(args.json, null, 2)}</pre>}
   </div>
 );