+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 6.2.0-beta.10 (March 5, 2021)
+
+### Bug Fixes
+
+- Angular: Keep story templates with an empty value ([#14113](https://github.com/storybookjs/storybook/pull/14113))
+- Core: Fix standalone API ... again ([#14140](https://github.com/storybookjs/storybook/pull/14140))
+
+## 6.2.0-beta.9 (March 4, 2021)
+
+### Bug Fixes
+
+- Core: Fix standalone API ([#14122](https://github.com/storybookjs/storybook/pull/14122))
+- Core: Fix main.ts/preview.ts handling ([#14123](https://github.com/storybookjs/storybook/pull/14123))
+
+## 6.2.0-beta.8 (March 4, 2021)
+
+### Features
+
+- Core: Add 'mapping' to support complex arg values ([#14100](https://github.com/storybookjs/storybook/pull/14100))
+
+## 6.2.0-beta.7 (March 4, 2021)
+
+Failed publish
+
 ## 6.1.21 (March 3, 2021)
 
 ### Bug Fixes